split up the mega metabot.unit.spec.tsx unit test file
