+++ enterprise/frontend/src/metabase-enterprise/metabot/tests/message.unit.spec.tsx
@@ -1,0 +1,79 @@
+import { screen } from "__support__/ui";
+
+import {
+  assertConversation,
+  createMockReadableStream,
+  createPauses,
+  enterChatMessage,
+  input,
+  mockAgentEndpoint,
+  responseLoader,
+  sendMessageButton,
+  setup,
+  whoIsYourFavoriteResponse,
+} from "./utils";
+
+describe("metabot > message", () => {
+  it("should properly send chat messages", async () => {
+    setup();
+
+    mockAgentEndpoint({
+      textChunks: whoIsYourFavoriteResponse,
+      initialDelay: 50,
+    });
+
+    await enterChatMessage("Who is your favorite?", false);
+    expect(await input()).toHaveTextContent("Who is your favorite?");
+
+    await enterChatMessage("Who is your favorite?");
+    expect(await responseLoader()).toBeInTheDocument();
+    expect(
+      await screen.findByText("You, but don't tell anyone."),
+    ).toBeInTheDocument();
+
+    expect(await input()).toHaveTextContent("");
+    expect(await input()).toHaveFocus();
+  });
+
+  it("should be able to send a message via send button", async () => {
+    setup();
+    mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
+
+    await enterChatMessage("Who is your favorite?", false);
+    expect(await input()).toHaveTextContent("Who is your favorite?");
+    (await sendMessageButton()).click();
+
+    expect(
+      await screen.findByText("You, but don't tell anyone."),
+    ).toBeInTheDocument();
+  });
+
+  it("should properly handle partial messages", async () => {
+    setup();
+
+    const [pause1] = createPauses(1);
+    mockAgentEndpoint({
+      stream: createMockReadableStream(
+        (async function* () {
+          yield `0:"You, but "\n`;
+          await pause1.promise;
+          yield `0:"don't tell anyone."\n`;
+          yield `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`;
+        })(),
+      ),
+    });
+
+    await enterChatMessage("Who is your favorite?");
+    await assertConversation([
+      ["user", "Who is your favorite?"],
+      ["agent", "You, but"],
+    ]);
+
+    pause1.resolve();
+
+    await assertConversation([
+      ["user", "Who is your favorite?"],
+      ["agent", "You, but don't tell anyone."],
+    ]);
+  });
+});
