+++ enterprise/frontend/src/metabase-enterprise/metabot/tests/utils.tsx
@@ -1,12 +1,44 @@
+import { combineReducers } from "@reduxjs/toolkit";
 import userEvent from "@testing-library/user-event";
+import fetchMock from "fetch-mock";
+import { assocIn } from "icepick";
 
-import { act, fireEvent, screen, waitFor } from "__support__/ui";
+import { setupEnterprisePlugins } from "__support__/enterprise";
+import { setupDatabaseListEndpoint } from "__support__/server-mocks";
+import { mockSettings } from "__support__/settings";
+import {
+  act,
+  fireEvent,
+  renderWithProviders,
+  screen,
+  waitFor,
+} from "__support__/ui";
 import {
   type MockStreamedEndpointParams,
+  createMockReadableStream,
+  createPauses,
   mockStreamedEndpoint,
 } from "metabase-enterprise/api/ai-streaming/test-utils";
+import type { User } from "metabase-types/api";
+import {
+  createMockTokenFeatures,
+  createMockUser,
+} from "metabase-types/api/mocks";
+import { createMockState } from "metabase-types/store/mocks";
 
-import { type MetabotAgentId, setVisible } from "../state";
+import { Metabot } from "../components/Metabot";
+import { FIXED_METABOT_IDS } from "../constants";
+import { MetabotProvider } from "../context";
+import {
+  type MetabotAgentId,
+  type MetabotState,
+  type MetabotStoreState,
+  metabotReducer,
+  setVisible,
+} from "../state";
+import { getMetabotInitialState } from "../state/reducer-utils";
+
+export { createMockReadableStream, createPauses };
 
 export const mockAgentEndpoint = (params: MockStreamedEndpointParams) =>
   mockStreamedEndpoint("/api/ee/metabot-v3/agent-streaming", params);
@@ -83,3 +115,79 @@ export const lastReqBody = async (
   await waitFor(() => expect(agentSpy).toHaveBeenCalled());
   return JSON.parse(agentSpy.mock.lastCall?.[1]?.body as string);
 };
+
+// Common mock response fixtures
+export const whoIsYourFavoriteResponse = [
+  `0:"You, but don't tell anyone."`,
+  `2:{"type":"state","version":1,"value":{"queries":{}}}`,
+  `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+];
+
+export const erroredResponse = [
+  `3:{"type": "error", "value": "Couldn't do the thing"}`,
+  `d:{"finishReason":"error","usage":{"promptTokens":4916,"completionTokens":8}}`,
+];
+
+// Setup function for metabot tests
+export function setup(
+  options: {
+    ui?: React.ReactElement;
+    metabotPluginInitialState?: MetabotState;
+    currentUser?: User | null | undefined;
+    promptSuggestions?: { prompt: string }[];
+  } | void,
+) {
+  const settings = mockSettings({
+    "token-features": createMockTokenFeatures({
+      metabot_v3: true,
+    }),
+  });
+
+  setupEnterprisePlugins();
+
+  const _metabotState = getMetabotInitialState();
+  const metabotState = assocIn(
+    _metabotState,
+    ["conversations", "omnibot", "visible"],
+    true,
+  );
+
+  const {
+    ui = <Metabot />,
+    currentUser = createMockUser(),
+    metabotPluginInitialState = metabotState,
+    promptSuggestions = [],
+  } = options || {};
+
+  fetchMock.get(
+    `path:/api/ee/metabot-v3/metabot/${FIXED_METABOT_IDS.DEFAULT}/prompt-suggestions`,
+    { prompts: promptSuggestions, offset: 0, limit: 3, total: 3 },
+  );
+  setupDatabaseListEndpoint([]);
+
+  const { store, rerender } = renderWithProviders(
+    <MetabotProvider>{ui}</MetabotProvider>,
+    {
+      storeInitialState: createMockState({
+        settings,
+        currentUser: currentUser ? currentUser : undefined,
+        plugins: {
+          metabotPlugin: metabotPluginInitialState,
+        },
+      } as any),
+      customReducers: {
+        plugins: combineReducers({
+          metabotPlugin: metabotReducer,
+        }),
+      },
+    },
+  );
+
+  return {
+    rerender,
+    conversationIds: Object.keys(metabotState.conversations),
+    store: store as Omit<typeof store, "getState"> & {
+      getState: () => MetabotStoreState;
+    },
+  };
+}
